hkml_list: Remove --open option
'--open' option is a workaround for slow list command.  Remove it and
try to make list command faster.  User can read the mails with context
on the list by using two windows, open list on one window, and open
content of specific mail of the list on the another window.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>