hkml_list: Use last listed mails for --thread option
For different filtering options, --thread works inconsistently.  E.g.,
--thread based on 'hkml list lkml --hot' works incorrectly.  That's
because '--thread' constructs mails to show from the scratch, which
results in index that different from the previous list.

As a workaround, ignore <source> and use last list for --thread option.
For a long term, it may makes sense to make --thread a separate index
dependent command.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>