list: Check if tag exists if source_type is None
Without this patch, if a mailing name has a typo it shows "no mail". Now
it prints an error:

$ hkml list live-patch
Mailing list 'live-patch' in manifest not found.

The correct ML name is live-patching. This patch preserves the current
behavior of hkml to guess what type of list the user is trying to do,
making the ML the last option to check. Without this patch, if the ML
didn't exists it should assume the user tried to search for a tag.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>