hkml_list: Use YYYY/MM/DD format for mails
The output of `hkml list` uses 'YY/MM/DD %H:%M' format for each mail's
date.  It was originally decided so to reduce length of the line.  But,
it turned out it only makes it confusing, and seems using two more
characters is not making output too long.  Change the format to use YYYY
instead.

Suggested-by: Marcos Paulo de Souza <mpdesouza@suse.com>
Closes: https://github.com/sjp38/hackermail/issues/5
Signed-off-by: SeongJae Park <sj@kernel.org>