hkml_thread: Support message-id input
Sometimes users would find mails from web or somewhere, and wants to
continue reading the full thread or the mail from hkml.  But 'hkml
thread' is currently receiving only the index of a mail in the thread
from the last-generated list.  Let the command to receive message id,
too.  Nonetheless, this commit implement it only for b4-use case.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>