src/hkml_view: Print each line only up to maximum column of the screen
If the a string to addstr() is longer than the columns of the screen,
curses prints whole string, even to next lines.  As a result, the next
line is not printed properly.  Fix the issue by checking the length of
the line and cutting it when it is longer than the screen.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>