TODO: Add an item for clean threads
With display effects menu, finding mails to focus on is easy.  E.g., Dim
mails older than specific day.  With that, --since option is only for
lighter overhead, and until when the threads parent mails should be
tracked.  Having yet another option to set the limit of until when
threads parent mails need to be tracked could be useful to reduce number
of splattered threads.

For example, 'hkml list damon --since 2024-10-15 --option_x 2024-10-10'
would show mails since 2024-10-15.  But, if a thread on the output has a
parent mail that sent before 2024-10-15, hkml will further fetch mails
since 2024-10-10, search the parent mail from it, and add it to the list
output if found.

Add a TODO item for that.

Signed-off-by: SeongJae Park <sj@kernel.org>