hkml_thread/get_thread_mails_from_web(): Support public inbox link URLs
For example, if we get a public inbox link URL from 'hkml monitor'
outputs and want to read it from 'hkml list', we need to get the message
id part from the url.  It is cumbersome.  Let 'hkml list' receive the
public inbox link url and parse the message id by itself.

Signed-off-by: SeongJae Park <sj@kernel.org>