_hkml_list_cache: save cache creation history in a separate file
Cache creation history is useful for --dim_old suggestion.  The history
is maintained in list output cache file, which contains list output, and
therefore have a restrict in the size (up to 64 cache items).  When old
cached list output is evicted, the cache creation history also goes.
Maintain a separate file for cache creation history.  This commit
implements saving part of it only.

Signed-off-by: SeongJae Park <sj@kernel.org>