hkml_view: Add {git, lore}-style coloring for text
Change color for text containing code additions to green, code deletions
to red, and original messages (when viewing a response) to blue.
Because code additions are now in green, focused lines are now in cyan
instead.

Also added a new item in TODO: creating a TUI for users to change
the colors of these 4 categories.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>
Signed-off-by: SeongJae Park <sj@kernel.org>