hkml_view: change focus line color to yello
With commit e88c710b294d ("hkml_view: Add {git, lore}-style coloring for
text"), the focus line color has changed from green to cyan, because
green is now for new lines on diff.  But cyan is not easy to distinguish
from dimmed white, which is the color for old mails on the mails list.
Use yellow instead.

Signed-off-by: SeongJae Park <sj@kernel.org>