hkml_view_mails: suggest dim_old before generating the list
Commit f629ff01f843 ("hkml_list: move suggest_dim_old() call to
hkml_view_mails") moved dim_old suggestion to be done after the list is
made.  Since the list creation dat is added together with the list
generation, the dim_old suggestion contains the option for just
generated date as the default.  As a result, everything is dimmed.  Fix
it by doing the suggestion before the generation of the list.

Fixes: f629ff01f843 ("hkml_list: move suggest_dim_old() call to hkml_view_mails")
Signed-off-by: SeongJae Park <sj@kernel.org>