hkml_view_mails: exclude <5 seconds old dim_old suggestion
hkml_view_mails create the list, and then suggest dim_old options.
Since the last creation dates are added when creating the list, the
latest option is just right now, and it is the default option.  As a
result, if users select the default, every mail on the list is dimmed.
Check the date of the latest suggestion, and skip it if it is less than
five seconds past.

Signed-off-by: SeongJae Park <sj@kernel.org>