hkml_view: rename scroll_cols of ScrollableList to focus_col
scroll_cols field of ScrollableList reperesents the amount of columns to
horizontally scroll.  It is actually also the position of the cursor,
and displayed as focused column.  Actually the field also represents the
focused column.  Rename the field to focus_col.  Also update the help
message of the horizontal focus moving keys (left/h/right/l).

Signed-off-by: SeongJae Park <sj@kernel.org>