hkml_send: use hkml_tag.handle_may_sent_mail()
The current logic suggests duplicate draft removal after trying sending
of a mail, only if the user wants to tag it as sent or a draft.  In some
situation, the user may not want to tag the sent or unsent mail, but
still want to cleanup duplicate drafts.  Support it by offloading all
questions to hkml_tag by using handle_may_sent_mail().

Signed-off-by: SeongJae Park <sj@kernel.org>