hkml_view_mails: remember and restore cursor position of mail contents
In common cases, users may open mail, read it, find something to check
from other mails on the list, go back to the list, read the other mail,
go back to the list again and open the original mail.  hkml reset the
cursor position to the beginning of the mail.  Remember and restore the
previous position of the cursor on the mail.

Signed-off-by: SeongJae Park <sj@kernel.org>