hkml_view_text: Handling Malformed Git Hunks
There are two cases for malformed git hunks. Either there are
more difflines than the header presents, or there are less
difflines than the header. In either case, we should stick to
a principle of coloring as many valid diff lines as we can,
without incorrectly coloring non-diff lines.

Case 1) len(diff) > header
In this case, we should end diff coloring when the header reports
the end of the diff. In some cases (as explained below), developers
might want to include partial difflogs to demonstrate a potential
change, without including the entire hunk. It is far less likely
that a developer will format a diff, and then add extra lines
after they have received the output.

Thus, in this case, we should still color all of the lines we
have already by returning the length of the hunk so far. This
changes old behavior, which would have returned -1 (and ignored
this hunk when coloring).

Case 2) len(diff) < header
In this case, we shoudl end diff coloring when there are no more
valid diff lines to read. That is, lines that do not begin with
{'+', '-', ' '}. Then, we can report the number of lines read
so far, as opposed to relying on just the header information.

This is the case that requires us to do the length incrementation
after having verified that the current line is a diffline. If
the current line is just normal text, we shouldn't include it as
part of the hunk, and return the number of difflines read so far.

It should also be noted in the patch that orig_content < 0 or
new_content < 0 should not happen, unless the header is malformed
in a way such that there are more total difflines and either:
(1) More + lines but less - lines than the header
(2) More - lines but less + lines than the header

Of course, this is a very rare situation. We do our best and report
the number of lines that we have read so far.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>