hkml_patch/format: add base commit's commit message to cover letter
In some workflows like that for DAMON maintainer, base commit is used as
a fake commit for marking start of a patch series that easy to find from
git log output.  Also the commit is usually be used to contain draft
cover letter messages.  Add the base commit's commit message to cover
letter, so that user can add and update the draft easily.

Signed-off-by: SeongJae Park <sj@kernel.org>