hkml_patch_format: only cc every get_maintainer.pl found recipients
The current mechanism adds get_maintainer.pl outputs except maling lists
as To:, and adds the mailing lists as Cc.  It is bit complicated, and
could confuse some people since it could give a nuance that To: people
are responsible to merge the patch.  Manual update of recipients is
therefore usually needed.  To make things simpler, just Cc all
recipients that automatically found.

Signed-off-by: SeongJae Park <sj@kernel.org>