hkml_view_mails: Fix view effect cancellation error
Currently, when a user opens the effect selection menu from
m --> 5: set display effects, and decide to not take any action and
press enter to exit out of the screen, hkml crashes. This is because
the effect will be set to None once entering the TUI, regardless
of whether a selection has been made. This temporary patch prevents
hkml from crashing when users follow this path. Another patch
will be sent out to address the root problem by not making
any changes to the effects if the user decides not to make any
changes.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>