hkml_view_mails: Fix effect cancellation
Currently, when a user opens the effect selection menu from
m --> 5: set display effects, and decide to not take any action and
press enter to exit out of the screen, hkml crashes. This is because
the effect will be set to None once entering the TUI, regardless
of whether a selection has been made. This patch prevents hkml from
crashing when users follow this path.

Note that commit ce96f983927f ("hkml_view_mails: Fix view effect
cancellation error"), which also fixes the crash and sent as a patch in
parallel to this commit's patch,  has already applied.  This commit is
for not only preventing the crash but also making the error handling
behavior better.

Moreover, a user will probably expect their old effects to be preserved
if they cancel during the effects setting menu. Currently, every time
a user enters the effect selection menu, a new MailDisplayEffect object
is created, therefore resetting their previous selections. This patch
preserves effects in between menu selections, so that if a user decides
to cancel the effect change, or enters an invalid selection, their
previous selction will be restored instead.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>
Signed-off-by: SeongJae Park <sj@kernel.org>