hkml_patch_format: add an option for adding maintainers of a file as common recipients
In some cases, all patches of the series should sent to single subsystem
maintainers so that the maintainer can add the patches to their tree.
But get_maintainer.pl fails finding the context for some of individual
files.

Add an option to specify that, namely --main_file.  If it is given, find
maitnainers of the file using get_maintainer.pl, and add them as common
Cc.

Signed-off-by: SeongJae Park <sj@kernel.org>