hkml_view: Allow users to enable search highlighting
Search highlighting is off by default, and users must opt-in to the
feature. The first time a user makes a search, a prompt asks them if
they would like to enable search highlighting (default is yes).

To avoid user fatigue, this prompt is only asked the first time a user
searches in an instance of hkml, and will set the option for the entirety
of the session. This patch does not offer a way for the user to change
their selection after it is made, meaning users must restart hkml if
they would like to change the option.

This patch also makes the default search keyword '[' on boot. That is,
users can now scroll through the list using n/N, and jump between mails,
rather than jumping between lines.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>
Signed-off-by: SeongJae Park <sj@kernel.org>