hkml_view: Improve search menu
This patch improves the search menu by indicating the current
status of the search function (current keyword, whether highlighting
is enabled or disabled). It also asks users every time for the
highlighting option, with the default still set to Y.

One other change this patch makes is that if a user does not
enter a search keyword, we can imagine that they do not want to
use the search function. In that case, we don't ask them if they
want to enable search highlighting.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>

Thanks,
SJ