hkml_thread: Don't crash with 0 arguments + no cache
The expected behavior of running 'hkml thread' with no message
identifier is to display the last cached thread. However, when there
is no cached thread, hkml crashes.

When there is no message identifier and no cached thread, inform the
user of what the problem is and exit gracefully instead of crashing.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>