hkml_list: Improve thread error handling
When expanding a thread fails (e.g. due to wget or gunzip
misconfiguration) and leads to a problem (e.g. failed to download mbox)
hkml provides no explanation for the failure.

We already keep track of the error message, so let's pass it to the user
so that they have more context about the failure.

Resolves: https://github.com/sjp38/hackermail/issues/12
Closes: https://github.com/sjp38/hackermail/issues/12
Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>
Signed-off-by: SeongJae Park <sj@kernel.org>