_hkml_list_cache: set cache creation date only if the list output is changed
Cache creation date, which is used as dim_old suggestion, is updated
from _hkml_list_cache.set_item() even if the content is not chaged.  As
a result, dim_old suggestion makes all the list output dimmed.  Avoid
updating the creation date if nothing has really changed on the list.

Signed-off-by: SeongJae Park <sj@kernel.org>