hkml_write: Hide 'which' subcommand output
When using "hkml write" or replying to a mail using the TUI,
hkml checks for what editors are installed on the user's system.
To do this, subprocess.check_output runs the "which" command
to see where the binary is located. If the binary is not
found, then which returns 1, which is caught by the try-except.

However, this error handling is not silent; users will see
that an error message is printed to their console saying that
the editor was not found. This is confusing, since users are
not expected to have all 4 editors installed. If a user only
has 1 editor installed, then they will see 3 error messages
printed to their output.

Instead of printing the "error" to stdout / TUI, ignore it,
since the error is handled within hkml anyways. Also, change
the description of the try-except block, since there are
other ways that this command can fail (e.g. binary for
"which" is not found)

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>
Signed-off-by: SeongJae Park <sj@kernel.org>