hkml_patch: support adding cover letter as a merge commit
Different subsystems have different ways to handle cover letter.  In
case of mm, it is added to the first patch.  In case of some other
subsystems including vfs, it is added as a fake merge commit.  The mm
way is supported by hkml now.  Let the merge commit way be supported.

Signed-off-by: SeongJae Park <sj@kernel.org>