hkml_write: Ignore hkml signature/body warning
In addition to the To: and Cc: hkml placeholders, let's also ignore the
"write above" and hkml signature warning lines.

Even though the To/Cc sections are necessary to send a proper mail
(if these aren't edited, then the mail will not send), I think that it is
reasonable that a user probably wouldn't want to send the write above
or hkml warning lines in their mail.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>
Signed-off-by: SeongJae Park <sj@kernel.org>