hkml_patch_format: support formatting patch series from topic branches
A common git workflow is topic branch.  In the workflow, some people and
subsystem makes a fake merge commit for managing additional information
about the topic.  For example, below log graph can be imagined.

    $ git log --graph --pretty="%s"
    *   Merge "topic A"
    |\
    | * topic A change 2
    | * topic A change 1
    |/
    * baseline

The merge commit is a good place to put cover letter content.  Make
'hkml patch format' to check if given commits to format as a patch is
made in the way, and add an option to use the merge commit as the cover
letter content source.

Signed-off-by: SeongJae Park <sj@kernel.org>