hkml_view: Handle non-chr characters
Some keyboard inputs are not in the valid chr() range for Python.
For instance, on Mac, resizing a screen using Cmd+ or Cmd- will cause
a crash in this code snippet, since x = self.screen.getch() will be -1.
In cases like this, chr(-1) will throw a ValueError and crash hkml.

It also doesn't make sense to even gracefully exit, since that can be
frustrating for users who have to restart their hkml instance every
time they try to resize their screen. Let's just silently ignore the
key press if hkml is unable to parse it.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Reviewed-by: SeongJae Park <sj@kernel.org>
Signed-off-by: SeongJae Park <sj@kernel.org>