hkml_tag: split tags file for 100 tag items
hkml_tag backs up tag data when tags file size exceeds 10 MiB, to avoid
sync failure due to large files be rejected by the remote git server.
The backup file is just ignored.

Simplify it by storing tags data into multiple files, with up to 100 tag
data items per file, in message id sorted order.  And let the logic read
all files and concatenate the items.

This is suboptimum.  But this might be sufficient for simple use cases.
Let's start with simple implementation and optimize later.

One risk of this change is that, if a user uses sync feature but using
different hkml versions on different machines, ther could be confusion.
The old version will download latest tags data, but read only a part of
the data.  Later syncing from it can even make data lost.  But, the
backup system is based on git, so we can revert the long change later.

Another risk of this change is that the .hkm repo size could increase.
But, again, it is just a git repo.  Users can squash the commits later.

Signed-off-by: SeongJae Park <sj@kernel.org>