tests/test_hkml_view_text: update malformed test expectation
The test text of test_hunk_lines_malformed_2() has an empty line at the
end.  Empty line was not treated as a part of hunk, but commit
3a79c525ca75 ("hkml_view_text/hunk_length(): support empty lines in
hunk") made it to be treated as a part of a hunk.  The changed behavior
is a correct behavior, but broke the test.  Fix it by updating the
expectation.

Fixes: 3a79c525ca75 ("hkml_view_text/hunk_length(): support empty lines in hunk")
Signed-off-by: SeongJae Park <sj@kernel.org>