hkml_open: Add extra line for context
Commit [ d5bf631a82a3ee8120a4b042cdb76464da91adbe hkml_view_text:
show mail context of the current line at the bottom ] added a very helpful
feature, which displays previous messages' origins when writing a repsonse.
However, to add this feature, a new line was created at the bottom of every
open mail instance. Sometimes, this can make reading the bottom of a mail
difficult, since the last line will sometimes be un-viewable due to the
context taking up an extra line.

To accomodate for this, let's unconditionally add an extra line at the
bottom of the mail list, so that the last line can always be rendered.
If the focused row is not a response, then the context line will be empty
and users will see an extra line at the bottom. If there is context,
then the empty line will be replaced by the context.

Signed-off-by: Joshua Hahn <joshua.hahnjy@gmail.com>
Signed-off-by: SeongJae Park <sj@kernel.org>