hkml_view: remove ScrollableList.last_drawn
At the beginning, hkml's interactive list was only for listing mails and
providing scrolls.  For replying or doing more actions, users had to use
other CLI commands like 'hkml reply' with the mail index argument.  To
let users easily see the index of the mails of the interest, the
interactive list was drawing the last shown things on the terminal
again.  Nowadays CLI commands like 'hkml reply' are nearly not being
used.  Remove the code for printing the last drawn things on the
terminal again.

Signed-off-by: SeongJae Park <sj@kernel.org>