hkml_view: add draw_fn and fn_args arguments for view()
The current implementation is obscure and hard-coded for some types.
Receiving a general function for generating and showing the content can
make it more general and easy to read what it does.  Add the function
and arguments to pass to the function for that.

Also make all arguments of view() has default values for future callers
not need to set old arguments that will not really be used.

Signed-off-by: SeongJae Park <sj@kernel.org>