hkml_view_mails: respect original decorator on refresh_list()
refresh_list() is called from mail_of_row() if the list is cached one,
and hence not having the line_nr_mail_map.  It resets decorator and set
uncollapsed.  Hence when a collapsed list is loaded from the cache and
the user tries to do something including manual --dim_old, the list is
automatically uncollapsed.  Fix the issue by respecting the original
decorator.

Signed-off-by: SeongJae Park <sj@kernel.org>