hkml_list: add --alias option for differentiating cached outputs
Users might want to have multiple sessions for same list creation
condition.

For example, a user may reading through a mails list.  If thre are so
many mails, the user may quit the list and open it again without --fetch
option, to continue reading it.  hkml_view_mails can read the cached
list output very fast, with restoration of the cursor position on the
list, so this is convenient.

Let's suppose a few of urgent mails are directly delivered to the user's
personal mail inbox (e.g., gmail).  Or, the user somehow got to know
some interesting mails sent to the mailing list.  If the user generate
the list using hkml, and if the generating options are same to the one
they generated the list that still reading through, the cached output
and te cursor position of the list is over-written.

To work around this problem, introduce a new option, --alias.  It does
nothing but just make the change to the outputs and cursor position
cache key with user-specified alias, so that users can keep multiple
sessions without desruptive overwrites of the cached output and status.

We introduced --no_output_cache for the same reason.  But it works for
only one more session, while --alias will work for any number of
sessions.  Following commit will revert --no_output_cache.

Signed-off-by: SeongJae Park <sj@kernel.org>