hkml_list: use hkml_dir relative path for Mail's gitdir
A git-based mail cache key is useless if the absolute path to the git
repo is different, since the cache key uses absolute path to the git
repo as a part of the key.  Use hkml_dir-relative path for the gitdir
field of Mail, so that future cache keys will have the hkml directory
relative git repo path.  _hkml.Mail.set_mbox() is the only place that
uses the field for invoking git commands, and it is already patched to
support both aboslute path and hkml_dir-relative path for the git repo.
So this change should be safe.

Signed-off-by: SeongJae Park <sj@kernel.org>