Revert "hkml_sync: add an experimental option to sync list outputs"
This reverts commit 4794507360ee1e7aa91036a5dec433be8d0a4540.

Let's suppose the option is once used, and the cached output is updated
but not yet synced.  Doing sync without the option will fail.  That
happens frequently, especially when sending mails from a list, for
tagging the mail as sent.

Also how to ensure the cached mails on a system is also available on
another system is not yet solved question.  It might just not worthy to
pursue.

For now, the feature is just confusing and broken.  Revert it.

Signed-off-by: SeongJae Park <sj@kernel.org>