hkml_view_mails: make thread dim_old suggestions based on all thread mails
--dim_old suggestion is made based on the list arguments.  Let's suppose
a thread is made once based on a mail's msgid and later the same thread
is made, but based on the msgid of another mail of the thread.  They
have different list arguments (different msgid), and hence the firstly
generated date is not considered as --dim_old candidaste for the second
list.  For thread mails list, find and use all creation dates for all
msgids of the thread.

Signed-off-by: SeongJae Park <sj@kernel.org>